Keep the app in the dock at all times while active.
