--- client/src/cordova/apple/xcode/ios/Outline/Classes/AppKitBridge/NSApplication+Outline.swift
@@ -1,36 +1,0 @@
-// Copyright 2025 The Outline Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import AppKit
-
-extension NSApplication {
-    func showMainWindow() {
-        mainMenu = MainMenu()
-        activate(ignoringOtherApps: true)
-        setActivationPolicy(.regular)
-        guard let mainWindow = getMainWindow() else {
-            return
-        }
-        mainWindow.makeKeyAndOrderFront(self)
-    }
-
-    func getMainWindow() -> NSWindow? {
-        for window in windows {
-            if String(describing: window).contains("UINSWindow") {
-                return window
-            }
-        }
-        return nil
-    }
-}
